+++ classes/Gems/Task/Import/CancelAppointmentsNotImportedTask.php
@@ -1,0 +1,46 @@
+<?php
+
+/**
+ *
+ * @package    Gems
+ * @subpackage Task\Import
+ * @author     Matijs de Jong <mjong@magnafacta.nl>
+ * @copyright  Copyright (c) 2022, Erasmus MC and MagnaFacta B.V.
+ * @license    New BSD License
+ */
+
+namespace Gems\Task\Import;
+
+/**
+ *
+ * @package    Gems
+ * @subpackage Task\Import
+ * @license    New BSD License
+ * @since      Class available since version 1.9.1
+ */
+class CancelAppointmentsNotImportedTask extends \MUtil_Task_TaskAbstract
+{
+    /**
+     * @var \Zend_Db_Adapter_Abstract
+     */
+    protected $db;
+    
+    /**
+     * @inheritDoc
+     */
+    public function execute($source = null, $cancelCode = null, $fromDate = null)
+    {
+        if ($source && $cancelCode && $fromDate) {
+            $values = ["gap_status" => $cancelCode];
+            
+            $where = $this->db->quoteInto("gap_source = ?", $source);
+            $where .= $this->db->quoteInto(" AND gap_status != ?", $cancelCode);
+            $where .= $this->db->quoteInto(" AND gap_last_synch < ?", $fromDate);
+
+            // \MUtil_Echo::track($values, $where);            
+            $deleted = $this->db->update('gems__appointments', $values, $where);
+            
+            $this->getBatch()->addMessage(sprintf($this->_('%s appointment(s) deleted.'), $deleted));
+        }
+    }
+}